# data/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/user/refresh": {
      "post": {
        "operationId": "UserAuthController_refresh",
        "summary": "modules.auth.user",
        "description": "Api refresh jwt accessToken",
        "parameters": [
          {
            "name": "user-agent",
            "in": "header",
            "description": "User agent header",
            "required": false,
            "schema": {
              "default": "Mozilla/5.0 (Macintosh; PPC Mac OS X 10_10_0 rv:4.0; DE) AppleWebKit/535.0.0 (KHTML, like Gecko) Version/6.1.6 Safari/535.0.0",
              "example": "Mozilla/5.0 (Macintosh; PPC Mac OS X 10_10_0 rv:4.0; DE) AppleWebKit/535.0.0 (KHTML, like Gecko) Version/6.1.6 Safari/535.0.0",
              "type": "string"
            }
          },
          {
            "name": "x-custom-lang",
            "in": "header",
            "description": "Custom language header",
            "required": false,
            "schema": { "default": "en", "example": "en", "type": "string" }
          },
          {
            "name": "x-timestamp",
            "in": "header",
            "description": "Timestamp header, in microseconds",
            "required": false,
            "schema": { "default": 1689049850358, "example": 1689049850358, "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                  "properties": {
                    "message": { "example": "user.refresh" },
                    "statusCode": { "type": "number", "example": 200 },
                    "data": { "$ref": "#/components/schemas/UserRefreshSerialization" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                      "properties": {
                        "message": { "example": "auth.error.refreshTokenUnauthorized" },
                        "statusCode": { "type": "number", "example": 5001 }
                      }
                    },
                    {
                      "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                      "properties": {
                        "message": { "example": "apiKey.error.keyNeeded" },
                        "statusCode": { "type": "number", "example": 5020 }
                      }
                    },
                    {
                      "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                      "properties": {
                        "message": { "example": "apiKey.error.notFound" },
                        "statusCode": { "type": "number", "example": 5021 }
                      }
                    },
                    {
                      "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                      "properties": {
                        "message": { "example": "apiKey.error.notActiveYet" },
                        "statusCode": { "type": "number", "example": 5023 }
                      }
                    },
                    {
                      "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                      "properties": {
                        "message": { "example": "apiKey.error.expired" },
                        "statusCode": { "type": "number", "example": 5024 }
                      }
                    },
                    {
                      "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                      "properties": {
                        "message": { "example": "apiKey.error.invalid" },
                        "statusCode": { "type": "number", "example": 5025 }
                      }
                    }
                  ]
                }
              }
            }
          },
          "408": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                  "properties": {
                    "message": { "example": "http.serverError.requestTimeout" },
                    "statusCode": { "type": "number", "example": 5052 }
                  }
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                  "properties": {
                    "message": { "example": "http.serverError.internalServerError" },
                    "statusCode": { "type": "number", "example": 5050 }
                  }
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                  "properties": {
                    "message": { "example": "http.serverError.serviceUnavailable" },
                    "statusCode": { "type": "number", "example": 5051 }
                  }
                }
              }
            }
          }
        },
        "tags": ["modules.auth.user"],
        "security": [{ "refreshToken": [] }, { "apiKey": [] }]
      }
    },
    "/api/v1/auth/user/change-password": {
      "patch": {
        "operationId": "UserAuthController_changePassword",
        "summary": "modules.auth.user",
        "description": "Api user change password",
        "parameters": [
          {
            "name": "user-agent",
            "in": "header",
            "description": "User agent header",
            "required": false,
            "schema": {
              "default": "Mozilla/5.0 (Windows; U; Windows NT 6.2)AppleWebKit/535.2.1 (KHTML, like Gecko) Version/5.0.2 Safari/535.2.1",
              "example": "Mozilla/5.0 (Windows; U; Windows NT 6.2)AppleWebKit/535.2.1 (KHTML, like Gecko) Version/5.0.2 Safari/535.2.1",
              "type": "string"
            }
          },
          {
            "name": "x-custom-lang",
            "in": "header",
            "description": "Custom language header",
            "required": false,
            "schema": { "default": "en", "example": "en", "type": "string" }
          },
          {
            "name": "x-timestamp",
            "in": "header",
            "description": "Timestamp header, in microseconds",
            "required": false,
            "schema": { "default": 1689049850359, "example": 1689049850359, "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserChangePasswordDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                  "properties": {
                    "message": { "example": "user.change-password" },
                    "statusCode": { "type": "number", "example": 200 }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                      "properties": {
                        "message": { "example": "user.error.passwordNotMatch" },
                        "statusCode": { "type": "number", "example": 5207 }
                      }
                    },
                    {
                      "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                      "properties": {
                        "message": { "example": "user.error.newPasswordMustDifference" },
                        "statusCode": { "type": "number", "example": 5208 }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                      "properties": {
                        "message": { "example": "auth.error.accessTokenUnauthorized" },
                        "statusCode": { "type": "number", "example": 5000 }
                      }
                    },
                    {
                      "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                      "properties": {
                        "message": { "example": "apiKey.error.keyNeeded" },
                        "statusCode": { "type": "number", "example": 5020 }
                      }
                    },
                    {
                      "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                      "properties": {
                        "message": { "example": "apiKey.error.notFound" },
                        "statusCode": { "type": "number", "example": 5021 }
                      }
                    },
                    {
                      "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                      "properties": {
                        "message": { "example": "apiKey.error.notActiveYet" },
                        "statusCode": { "type": "number", "example": 5023 }
                      }
                    },
                    {
                      "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                      "properties": {
                        "message": { "example": "apiKey.error.expired" },
                        "statusCode": { "type": "number", "example": 5024 }
                      }
                    },
                    {
                      "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                      "properties": {
                        "message": { "example": "apiKey.error.invalid" },
                        "statusCode": { "type": "number", "example": 5025 }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                  "properties": {
                    "message": { "example": "user.error.passwordAttemptMax" },
                    "statusCode": { "type": "number", "example": 5210 }
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                  "properties": {
                    "message": { "example": "user.error.notFound" },
                    "statusCode": { "type": "number", "example": 5200 }
                  }
                }
              }
            }
          },
          "408": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                  "properties": {
                    "message": { "example": "http.serverError.requestTimeout" },
                    "statusCode": { "type": "number", "example": 5052 }
                  }
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                  "properties": {
                    "message": { "example": "http.serverError.internalServerError" },
                    "statusCode": { "type": "number", "example": 5050 }
                  }
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                  "properties": {
                    "message": { "example": "http.serverError.serviceUnavailable" },
                    "statusCode": { "type": "number", "example": 5051 }
                  }
                }
              }
            }
          }
        },
        "tags": ["modules.auth.user"],
        "security": [{ "accessToken": [] }, { "apiKey": [] }]
      }
    },
    "/api/v1/auth/user/info": {
      "get": {
        "operationId": "UserAuthController_info",
        "summary": "modules.auth.user",
        "description": "Api user serialization information",
        "parameters": [
          {
            "name": "user-agent",
            "in": "header",
            "description": "User agent header",
            "required": false,
            "schema": {
              "default": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_5_9)  AppleWebKit/538.2.1 (KHTML, like Gecko) Chrome/24.0.898.0 Safari/538.2.1",
              "example": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_5_9)  AppleWebKit/538.2.1 (KHTML, like Gecko) Chrome/24.0.898.0 Safari/538.2.1",
              "type": "string"
            }
          },
          {
            "name": "x-custom-lang",
            "in": "header",
            "description": "Custom language header",
            "required": false,
            "schema": { "default": "en", "example": "en", "type": "string" }
          },
          {
            "name": "x-timestamp",
            "in": "header",
            "description": "Timestamp header, in microseconds",
            "required": false,
            "schema": { "default": 1689049850359, "example": 1689049850359, "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                  "properties": {
                    "message": { "example": "user.info" },
                    "statusCode": { "type": "number", "example": 200 },
                    "data": { "$ref": "#/components/schemas/UserInfoSerialization" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                      "properties": {
                        "message": { "example": "auth.error.accessTokenUnauthorized" },
                        "statusCode": { "type": "number", "example": 5000 }
                      }
                    },
                    {
                      "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                      "properties": {
                        "message": { "example": "apiKey.error.keyNeeded" },
                        "statusCode": { "type": "number", "example": 5020 }
                      }
                    },
                    {
                      "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                      "properties": {
                        "message": { "example": "apiKey.error.notFound" },
                        "statusCode": { "type": "number", "example": 5021 }
                      }
                    },
                    {
                      "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                      "properties": {
                        "message": { "example": "apiKey.error.notActiveYet" },
                        "statusCode": { "type": "number", "example": 5023 }
                      }
                    },
                    {
                      "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                      "properties": {
                        "message": { "example": "apiKey.error.expired" },
                        "statusCode": { "type": "number", "example": 5024 }
                      }
                    },
                    {
                      "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                      "properties": {
                        "message": { "example": "apiKey.error.invalid" },
                        "statusCode": { "type": "number", "example": 5025 }
                      }
                    }
                  ]
                }
              }
            }
          },
          "408": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                  "properties": {
                    "message": { "example": "http.serverError.requestTimeout" },
                    "statusCode": { "type": "number", "example": 5052 }
                  }
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                  "properties": {
                    "message": { "example": "http.serverError.internalServerError" },
                    "statusCode": { "type": "number", "example": 5050 }
                  }
                }
              }
            }
          },
          "503": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [{ "$ref": "#/components/schemas/ResponseDefaultSerialization" }],
                  "properties": {
                    "message": { "example": "http.serverError.serviceUnavailable" },
                    "statusCode": { "type": "number", "example": 5051 }
                  }
                }
              }
            }
          }
        },
        "tags": ["modules.auth.user"],
        "security": [{ "accessToken": [] }, { "apiKey": [] }]
      }
    },
...
  },
  "info": {
    "title": "SpecFinderAPI APIs Specification",
    "description": "Section for describe whole APIs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [{ "url": "/" }, { "url": "/staging" }, { "url": "/prod" }],
  "components": {
    "securitySchemes": {
      "accessToken": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" },
      "refreshToken": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" },
      "apiKey": { "type": "apiKey", "in": "header", "name": "x-api-key" }
    },
    "schemas": {
      "ResponseDefaultSerialization": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "nullable": false,
            "description": "return specific status code for every endpoints",
            "example": 200
          },
          "message": {
            "nullable": false,
            "description": "Message base on language",
            "oneOf": [
              { "type": "string", "example": "message endpoint" },
              {
                "type": "object",
                "example": { "en": "This is test endpoint.", "id": "Ini adalah endpoint test" }
              }
            ]
          },
          "_metadata": {
            "type": "object",
            "nullable": false,
            "description": "Contain metadata about API",
            "example": {
              "languages": ["en"],
              "timestamp": 1660190937231,
              "timezone": "Asia/Jakarta",
              "requestId": "40c2f734-7247-472b-bc26-8eff6e669781",
              "path": "/api/v1/test/hello",
              "version": "1",
              "repoVersion": "1.0.0"
            }
          }
        },
        "required": ["statusCode", "message", "_metadata"]
      },
...
      "UserRefreshSerialization": {
        "type": "object",
        "properties": {
          "tokenType": { "type": "string", "example": "Bearer", "nullable": false },
          "expiresIn": {
            "type": "string",
            "example": 1660190937231,
            "description": "Expire in timestamp",
            "nullable": false
          },
          "accessToken": {
            "type": "string",
            "example": "20hU8c43hjmdl3j7IdkCRh0CFGm6Gc",
            "description": "Will be valid JWT Encode string",
            "nullable": false
          },
          "refreshToken": {
            "type": "string",
            "example": "gpJbsJMEBMYdNs0xlIMwo6itgeuirg",
            "description": "Will be valid JWT Encode string",
            "nullable": false
          }
        },
        "required": ["tokenType", "expiresIn", "accessToken", "refreshToken"]
      },
      "UserChangePasswordDto": {
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string",
            "description": "new string password, newPassword can't same with oldPassword",
            "example": "hahu7",
            "nullable": false
          },
          "oldPassword": {
            "type": "string",
            "description": "old string password",
            "example": "n9ens",
            "nullable": false
          }
        },
        "required": ["newPassword", "oldPassword"]
      },
      "UserPayloadPermissionSerialization": {
        "type": "object",
        "properties": {
          "subject": {
            "type": "string",
            "nullable": false,
            "enum": ["API_KEY", "SETTING", "ROLE", "USER"],
            "example": "SETTING"
          },
          "action": { "type": "string", "nullable": false, "example": "1,3" }
        },
        "required": ["subject", "action"]
      },
      "UserInfoSerialization": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Id that representative with your target data",
            "example": "631d9f32a65cf07250b8938c",
            "nullable": false
          },
          "username": { "type": "string", "example": "Neil.Carter53", "nullable": true },
          "email": {
            "type": "string",
            "nullable": false,
            "example": "Maynard_Stoltenberg79@gmail.com"
          },
          "mobileNumber": { "type": "string", "nullable": true, "example": "8780817382" },
          "firstName": { "type": "string", "nullable": false, "example": "Jaquan" },
          "lastName": { "type": "string", "nullable": false, "example": "Braun-Paucek" },
          "role": {
            "type": "string",
            "example": "3c36af2a-c725-447b-892c-8b94864b0cd1",
            "nullable": false
          },
          "type": {
            "type": "string",
            "example": "ADMIN",
            "enum": ["SUPER_ADMIN", "USER", "ADMIN"],
            "nullable": false
          },
          "permissions": {
            "nullable": false,
            "example": [{ "subject": "API_KEY", "acton": "1,3" }],
            "type": "array",
            "items": { "$ref": "#/components/schemas/UserPayloadPermissionSerialization" }
          }
        },
        "required": ["_id", "email", "firstName", "lastName", "role", "type", "permissions"]
      },
...
    }
  }
}